# shared/api/swagger.yaml
basePath: /
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  piemdm_internal_model.ApprovalDefinition:
    properties:
      code:
        description: 审批定义唯一编码
        type: string
      createdAt:
        description: 创建时间
        type: string
      createdBy:
        description: 审计字段
        type: string
      deletedAt:
        allOf:
        - $ref: '#/definitions/gorm.DeletedAt'
        description: 删除时间
      description:
        description: 审批描述
        maxLength: 500
        type: string
      formData:
        description: |-
          表单配置
          FormConfig string `gorm:"type:text"` // 表单配置JSON
          FormSchema string `gorm:"type:text"` // 表单结构定义
        type: string
      id:
        type: integer
      name:
        description: 审批名称
        maxLength: 128
        type: string
      nodeList:
        description: |-
          流程配置
          ProcessConfig string `gorm:"type:text"` // 流程配置JSON
          ProcessNodes  string `gorm:"type:text"` // 流程节点配置
        type: string
      platform:
        description: |-
          状态和元数据，
          Builtin 系统内置,Custom 自定义,Feishu, DingTalk, WeChatWork
        type: string
      status:
        description: 状态:Normal 正常, Frozen 已冻结, Deleted 已删除
        type: string
      updatedAt:
        description: 更新时间
        type: string
      updatedBy:
        description: 更新人
        type: string
    required:
    - name
    type: object
...
  piemdm_internal_transport_request.CreateApprovalDefRequest:
    properties:
      code:
        maxLength: 128
        type: string
      description:
        description: |-
          EntityCode  string `json:"entityCode" binding:"required,min=1,max=64" comment:"关联实体编码"`
          Category    string `json:"category" binding:"max=64" comment:"审批分类"`
        maxLength: 500
        type: string
      formData:
        maxLength: 10000
        type: string
      name:
        maxLength: 128
        minLength: 1
        type: string
      nodeList:
        maxLength: 10000
        type: string
      platform:
        enum:
        - Builtin
        - Feishu
        - DingTalk
        - WeChatWork
        - Custom
        type: string
      status:
        enum:
        - Normal
        - Frozen
        - Deleted
        type: string
    required:
    - name
    - platform
    type: object
...
  piemdm_internal_transport_request.CreateVersionRequest:
    properties:
      comment:
        maxLength: 255
        minLength: 1
        type: string
      id:
        type: integer
    required:
    - comment
    - id
    type: object
...
  piemdm_internal_transport_request.UpdateApprovalDefRequest:
    properties:
      description:
        description: |-
          EntityCode  string `json:"entityCode" binding:"required,min=1,max=64" comment:"关联实体编码"`
          Category    string `json:"category" binding:"max=64" comment:"审批分类"`
        maxLength: 500
        type: string
      formData:
        maxLength: 10000
        type: string
      id:
        type: integer
      name:
        maxLength: 128
        minLength: 1
        type: string
      nodeList:
        maxLength: 10000
        type: string
      platform:
        enum:
        - Builtin
        - Feishu
        - DingTalk
        - WeChatWork
        - Custom
        type: string
      status:
        enum:
        - Normal
        - Frozen
        - Deleted
        type: string
    required:
    - id
    - name
    - platform
    type: object
...
host: localhost:8787
info:
  contact:
    email: support@piemdm.com
    name: API Support
  description: PieMDM Master Data Management System API Documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: PieMDM API
  version: "1.0"
paths:
...
  /api/v1/approval-defs:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        maxLength: 128
        name: code
        type: string
      - in: query
        name: endDate
        type: string
      - in: query
        maxLength: 128
        name: name
        type: string
      - in: query
        name: page
        type: integer
      - in: query
        name: pageSize
        type: integer
      - in: query
        name: startDate
        type: string
      - description: |-
          EntityCode string `form:"entityCode" binding:"max=64" comment:"实体编码"`
          Category   string `form:"category" binding:"max=64" comment:"分类"`
        enum:
        - Normal
        - Frozen
        - Deleted
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/piemdm_internal_model.ApprovalDefinition'
            type: array
      summary: 获取审批定义列表
      tags:
      - 审批定义
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建审批定义请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/piemdm_internal_transport_request.CreateApprovalDefRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/piemdm_internal_model.ApprovalDefinition'
      summary: 创建审批定义
      tags:
      - 审批定义
  /api/v1/approval-defs/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 审批定义ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 删除审批定义
      tags:
      - 审批定义
    get:
      consumes:
      - application/json
      parameters:
      - description: 审批定义ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/piemdm_internal_model.ApprovalDefinition'
      summary: 获取审批定义详情
      tags:
      - 审批定义
    put:
      consumes:
      - application/json
      parameters:
      - description: 审批定义ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新审批定义请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/piemdm_internal_transport_request.UpdateApprovalDefRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/piemdm_internal_model.ApprovalDefinition'
      summary: 更新审批定义
      tags:
      - 审批定义
  /api/v1/approval-defs/{id}/activate:
    post:
      consumes:
      - application/json
      parameters:
      - description: 审批定义ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 激活审批定义
      tags:
      - 审批定义
  /api/v1/approval-defs/{id}/deactivate:
    post:
      consumes:
      - application/json
      parameters:
      - description: 审批定义ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 停用审批定义
      tags:
      - 审批定义
  /api/v1/approval-defs/{id}/publish:
    post:
      consumes:
      - application/json
      parameters:
      - description: 审批定义ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 发布审批定义
      tags:
      - 审批定义
  /api/v1/approval-defs/{id}/validate:
    post:
      consumes:
      - application/json
      parameters:
      - description: 审批定义ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 验证审批定义
      tags:
      - 审批定义
  /api/v1/approval-defs/{id}/versions:
    post:
      consumes:
      - application/json
      parameters:
      - description: 审批定义ID
        in: path
        name: id
        required: true
        type: integer
      - description: 创建版本请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/piemdm_internal_transport_request.CreateVersionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/piemdm_internal_model.ApprovalDefinition'
      summary: 创建审批定义新版本
      tags:
      - 审批定义
  /api/v1/approval-defs/code/{code}/versions:
    get:
      consumes:
      - application/json
      parameters:
      - description: 审批定义编码
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/piemdm_internal_model.ApprovalDefinition'
            type: array
      summary: 获取审批定义版本列表
      tags:
      - 审批定义
  /api/v1/approval-defs/entity/{entityCode}:
    get:
      consumes:
      - application/json
      parameters:
      - description: 实体编码
        in: path
        name: entityCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/piemdm_internal_model.ApprovalDefinition'
            type: array
      summary: 根据实体编码获取审批定义列表
      tags:
      - 审批定义
  /api/v1/approval-defs/entity/{entityCode}/active:
    get:
      consumes:
      - application/json
      parameters:
      - description: 实体编码
        in: path
        name: entityCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/piemdm_internal_model.ApprovalDefinition'
            type: array
      summary: 根据实体编码获取激活的审批定义列表
      tags:
      - 审批定义
  /api/v1/approval-defs/sync-feishu:
    post:
      consumes:
      - application/json
      parameters:
      - description: 审批定义编码 (飞书中的 Approval Code)
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 同步飞书审批定义 (表单)
      tags:
      - 审批定义
...
  /api/v1/approval_defs/code/{code}:
    get:
      consumes:
      - application/json
      parameters:
      - description: 审批定义编码
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/piemdm_internal_model.ApprovalDefinition'
      summary: 根据编码获取审批定义
      tags:
      - 审批定义
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"